Forgot to modify data type for the return value (double)

+++ src/palmapper/QPalma.h
@@ -128,7 +128,7 @@ public:
 	int capture_hits_2(Hits &hits, Result &result, bool non_consensus_search=false) const;
 	int perform_alignment(Result &result, Hits &readMappings, std::string &read_string, std::string &read_quality, std::string &dna, std::vector<region_t *> &regions, std::vector<int> &positions,
 						  Chromosome const &contig_id, char strand, int ori, int hit_read, int hit_dna, int hit_length, bool non_consensus_search, ALIGNMENT *& aln) const;
-	float score_unspliced(Read const &read, const char * read_anno, const char strand, const char ori) const;
+	double score_unspliced(Read const &read, const char * read_anno, const char strand, const char ori) const;
 	//void capture_hits_timing(int read_count=-1, float this_read=-1.0) const;
 	
 protected:
